ingest/requirements.txt
Flask==2.3.2
google-cloud-bigquery==3.17.2